Refactored Harmony patching
